chore:  types파일 제거
