+++ frontend/src/components/FeedbackText/FeedbackText.styles.ts
@@ -3,4 +3,6 @@ import { css } from '@emotion/react';
 
 export const feedbackText = (theme: Theme) => css`
   ${theme.typography.inter.small}
+
+  line-height: 20px;
 `;
