+++ frontend/src/components/FeedbackText/FeedbackText.tsx
@@ -1,6 +1,9 @@
 import { useAppTheme } from '@/hooks/useAppTheme';
-import { FeedbackTextProps } from './FeedbackText.types';
 import { feedbackText } from './FeedbackText.styles';
+
+export interface FeedbackTextProps {
+  text: string;
+}
 
 export default function FeedbackText({ text }: FeedbackTextProps) {
   const theme = useAppTheme();
