--- frontend/src/components/FeedbackText/FeedbackText.types.ts
@@ -1,3 +1,0 @@
-export interface FeedbackTextProps {
-  text: string;
-}
