--- frontend/src/components/Tag/Tag.types.ts
@@ -1,6 +1,0 @@
-import { SerializedStyles } from '@emotion/react';
-
-export interface TagProps {
-  children: React.ReactNode;
-  customCss?: SerializedStyles;
-}
