--- frontend/src/domains/user/home/components/CategoryTag/CategoryTag.types.ts
@@ -1,6 +1,0 @@
-export interface CategoryTagProps {
-  text: string;
-  type: 'complete' | 'incomplete';
-}
-
-export type CategoryType = 'complete' | 'incomplete';
