--- frontend/src/domains/user/home/components/StatusTag/StatusTag.types.ts
@@ -1,6 +1,0 @@
-export interface StatusTagProps {
-  children: React.ReactNode;
-  type: StatusType;
-}
-
-export type StatusType = 'complete' | 'incomplete';
